+++ lib/msal-browser/test/app/PublicClientApplication.spec.ts
@@ -6,7 +6,7 @@
 import sinon from "sinon";
 import { PublicClientApplication } from "../../src/app/PublicClientApplication";
 import { TEST_CONFIG, TEST_URIS, TEST_HASHES, TEST_TOKENS, TEST_DATA_CLIENT_INFO, TEST_TOKEN_LIFETIMES, RANDOM_TEST_GUID, DEFAULT_OPENID_CONFIG_RESPONSE, testNavUrl, testLogoutUrl, TEST_STATE_VALUES, testNavUrlNoRequest, DEFAULT_TENANT_DISCOVERY_RESPONSE } from "../utils/StringConstants";
-import { ServerError, Constants, AccountInfo, TokenClaims, PromptValue, AuthenticationResult, CommonAuthorizationCodeRequest, CommonAuthorizationUrlRequest, AuthToken, PersistentCacheKeys, AuthorizationCodeClient, ResponseMode, AccountEntity, ProtocolUtils, AuthenticationScheme, RefreshTokenClient, Logger, ServerTelemetryEntity, CommonSilentFlowRequest, CommonEndSessionRequest, LogLevel, NetworkResponse, ServerAuthorizationTokenResponse } from "@azure/msal-common";
+import { ServerError, Constants, AccountInfo, TokenClaims, PromptValue, AuthenticationResult, CommonAuthorizationCodeRequest, CommonAuthorizationUrlRequest, AuthToken, PersistentCacheKeys, AuthorizationCodeClient, ResponseMode, AccountEntity, ProtocolUtils, AuthenticationScheme, RefreshTokenClient, Logger, ServerTelemetryEntity, CommonSilentFlowRequest, CommonEndSessionRequest, LogLevel, NetworkResponse, ServerAuthorizationTokenResponse, CcsCredential, CcsCredentialType } from "@azure/msal-common";
 import { BrowserUtils } from "../../src/utils/BrowserUtils";
 import { BrowserConstants, TemporaryCacheKeys, ApiId, InteractionType, BrowserCacheLocation, WrapperSKU } from "../../src/utils/BrowserConstants";
 import { Base64Encode } from "../../src/encode/Base64Encode";
@@ -26,6 +26,7 @@ import { NavigationOptions } from "../../src/navigation/NavigationOptions";
 import { PopupUtils } from "../../src/utils/PopupUtils";
 import { EndSessionPopupRequest } from "../../src/request/EndSessionPopupRequest";
 import { EventMessage } from "../../src/event/EventMessage";
+import { EventHandler } from "../../src/event/EventHandler";
 
 describe("PublicClientApplication.ts Class Unit Tests", () => {
     const cacheConfig = {
@@ -67,7 +68,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
     describe("Constructor tests", () => {
 
         it("passes null check", (done) => {
-            expect(pca).not.toBeNull;
+            expect(pca).not.toBe(null);
             expect(pca instanceof PublicClientApplication).toBeTruthy();
             done();
         });
@@ -77,7 +78,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 sinon.stub(pca, <any>"interactionInProgress").returns(false);
                 window.location.hash = TEST_HASHES.TEST_SUCCESS_CODE_HASH_REDIRECT;
                 window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.ORIGIN_URI}`, TEST_URIS.TEST_ALTERNATE_REDIR_URI);
-                expect(await pca.handleRedirectPromise()).toBeNull;
+                expect(await pca.handleRedirectPromise()).toBe(null);
             }
         );
 
@@ -337,8 +338,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 expect(window.sessionStorage.length).toEqual(4);
             });
 
-            it(
-                "Multiple concurrent calls to handleRedirectPromise return the same promise",
+            it("Multiple concurrent calls to handleRedirectPromise return the same promise",
                 async () => {
                     const b64Encode = new Base64Encode();
                     const stateString = TEST_STATE_VALUES.TEST_STATE_REDIRECT;
@@ -424,7 +424,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                     const tokenResponse1 = await promise1;
                     const tokenResponse2 = await promise2;
                     const tokenResponse3 = await pca.handleRedirectPromise("testHash");
-                    expect(tokenResponse3).toBeNull();
+                    expect(tokenResponse3).toBe(null);
                     const tokenResponse4 = await pca.handleRedirectPromise();
 
                     if (!tokenResponse1 || !tokenResponse2) {
@@ -589,8 +589,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 }
             );
 
-            it(
-                "calls custom navigateInternal function then processes hash",
+            it("calls custom navigateInternal function then processes hash",
                 async () => {
                     const b64Encode = new Base64Encode();
                     const stateString = TEST_STATE_VALUES.TEST_STATE_REDIRECT;
@@ -690,7 +689,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
                     const tokenResponse = await pca.handleRedirectPromise();
                     if (!tokenResponse) {
-                        expect(tokenResponse).not.toBeNull();
+                        expect(tokenResponse).not.toBe(null);
                         throw new Error("Token Response is null!"); // Throw to resolve Typescript complaints below
                     }
                     expect(callbackCalled).toBeTruthy();
@@ -706,8 +705,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 }
             );
 
-            it(
-                "processes hash if navigateToLoginRequestUri is false and request origin is different",
+            it("processes hash if navigateToLoginRequestUri is false and request origin is different",
                 async () => {
                     const b64Encode = new Base64Encode();
                     const stateString = TEST_STATE_VALUES.TEST_STATE_REDIRECT;
@@ -811,8 +809,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
         describe("loginRedirect", () => {
 
-            it(
-                "loginRedirect throws an error if interaction is currently in progress",
+            it("loginRedirect throws an error if interaction is currently in progress",
                 async () => {
                     window.sessionStorage.setItem(`${Constants.CACHE_PREFIX}.${TEST_CONFIG.MSAL_CLIENT_ID}.${TemporaryCacheKeys.INTERACTION_STATUS_KEY}`, BrowserConstants.INTERACTION_IN_PROGRESS_VALUE);
                     // @ts-ignore
@@ -859,8 +856,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 pca.loginRedirect(loginRequest);
             });
 
-            it(
-                "loginRedirect navigates to created login url, with empty request",
+            it("loginRedirect navigates to created login url, with empty request",
                 (done) => {
                     sinon.stub(RedirectHandler.prototype, "initiateAuthRequest").callsFake((navigateUrl): Promise<void> => {
                         expect(navigateUrl.startsWith(testNavUrlNoRequest)).toBeTruthy();
@@ -908,6 +904,111 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 expect(browserStorage.getTemporaryCache(browserStorage.generateStateKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(TEST_STATE_VALUES.TEST_STATE_REDIRECT);
                 expect(browserStorage.getTemporaryCache(browserStorage.generateNonceKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(RANDOM_TEST_GUID);
                 expect(browserStorage.getTemporaryCache(browserStorage.generateAuthorityKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(`${Constants.DEFAULT_AUTHORITY}`);
+            });
+
+            it("Adds login_hint as CCS cache entry to the cache and urlNavigate", async () => {
+                const testIdTokenClaims: TokenClaims = {
+                    "ver": "2.0",
+                    "iss": "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0",
+                    "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
+                    "name": "Abe Lincoln",
+                    "preferred_username": "AbeLi@microsoft.com",
+                    "oid": "00000000-0000-0000-66f3-3332eca7ea81",
+                    "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
+                    "nonce": "123523",
+                };
+                const testCcsCred: CcsCredential = {
+                    credential: testIdTokenClaims.preferred_username || "",
+                    type: CcsCredentialType.UPN
+                };
+                const emptyRequest: CommonAuthorizationUrlRequest = {
+                    redirectUri: TEST_URIS.TEST_REDIR_URI,
+                    scopes: [],
+                    state: TEST_STATE_VALUES.USER_STATE,
+                    authority: TEST_CONFIG.validAuthority,
+                    correlationId: TEST_CONFIG.CORRELATION_ID,
+                    responseMode: TEST_CONFIG.RESPONSE_MODE as ResponseMode,
+                    nonce: "",
+                    authenticationScheme: TEST_CONFIG.TOKEN_TYPE_BEARER as AuthenticationScheme,
+                    loginHint: testIdTokenClaims.preferred_username || ""
+                };
+
+                sinon.stub(CryptoOps.prototype, "generatePkceCodes").resolves({
+                    challenge: TEST_CONFIG.TEST_CHALLENGE,
+                    verifier: TEST_CONFIG.TEST_VERIFIER
+                });
+
+                sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+                sinon.stub(NavigationClient.prototype, "navigateExternal").callsFake((urlNavigate: string, options: NavigationOptions): Promise<boolean> => {
+                    expect(options.noHistory).toBeFalsy();
+                    expect(urlNavigate).not.toBe("");
+                    return Promise.resolve(true);
+                });
+                const testLogger = new Logger(loggerOptions);
+
+                const browserCrypto = new CryptoOps();
+                const browserStorage = new BrowserCacheManager(TEST_CONFIG.MSAL_CLIENT_ID, cacheConfig, browserCrypto, testLogger);
+                await pca.loginRedirect(emptyRequest);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateStateKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(TEST_STATE_VALUES.TEST_STATE_REDIRECT);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateNonceKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(RANDOM_TEST_GUID);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateAuthorityKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(`${Constants.DEFAULT_AUTHORITY}`);
+                expect(browserStorage.getTemporaryCache(TemporaryCacheKeys.CCS_CREDENTIAL, true)).toEqual(JSON.stringify(testCcsCred));
+            });
+
+            it("Adds account homeAccountId as CCS cache entry to the cache and urlNavigate", async () => {
+                const testIdTokenClaims: TokenClaims = {
+                    "ver": "2.0",
+                    "iss": "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0",
+                    "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
+                    "name": "Abe Lincoln",
+                    "preferred_username": "AbeLi@microsoft.com",
+                    "oid": "00000000-0000-0000-66f3-3332eca7ea81",
+                    "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
+                    "nonce": "123523",
+                };
+                const testAccount: AccountInfo = {
+                    homeAccountId: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
+                    localAccountId: TEST_DATA_CLIENT_INFO.TEST_UID,
+                    environment: "login.windows.net",
+                    tenantId: testIdTokenClaims.tid || "",
+                    username: testIdTokenClaims.preferred_username || ""
+                };
+                const testCcsCred: CcsCredential = {
+                    credential: testAccount.homeAccountId,
+                    type: CcsCredentialType.HOME_ACCOUNT_ID
+                };
+                const emptyRequest: CommonAuthorizationUrlRequest = {
+                    redirectUri: TEST_URIS.TEST_REDIR_URI,
+                    scopes: [],
+                    state: TEST_STATE_VALUES.USER_STATE,
+                    authority: TEST_CONFIG.validAuthority,
+                    correlationId: TEST_CONFIG.CORRELATION_ID,
+                    responseMode: TEST_CONFIG.RESPONSE_MODE as ResponseMode,
+                    nonce: "",
+                    authenticationScheme: TEST_CONFIG.TOKEN_TYPE_BEARER as AuthenticationScheme,
+                    account: testAccount
+                };
+
+                sinon.stub(CryptoOps.prototype, "generatePkceCodes").resolves({
+                    challenge: TEST_CONFIG.TEST_CHALLENGE,
+                    verifier: TEST_CONFIG.TEST_VERIFIER
+                });
+
+                sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+                sinon.stub(NavigationClient.prototype, "navigateExternal").callsFake((urlNavigate: string, options: NavigationOptions): Promise<boolean> => {
+                    expect(options.noHistory).toBeFalsy();
+                    expect(urlNavigate).not.toBe("");
+                    return Promise.resolve(true);
+                });
+                const testLogger = new Logger(loggerOptions);
+
+                const browserCrypto = new CryptoOps();
+                const browserStorage = new BrowserCacheManager(TEST_CONFIG.MSAL_CLIENT_ID, cacheConfig, browserCrypto, testLogger);
+                await pca.loginRedirect(emptyRequest);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateStateKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(TEST_STATE_VALUES.TEST_STATE_REDIRECT);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateNonceKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(RANDOM_TEST_GUID);
+                expect(browserStorage.getTemporaryCache(browserStorage.generateAuthorityKey(TEST_STATE_VALUES.TEST_STATE_REDIRECT))).toEqual(`${Constants.DEFAULT_AUTHORITY}`);
+                expect(browserStorage.getTemporaryCache(TemporaryCacheKeys.CCS_CREDENTIAL, true)).toEqual(JSON.stringify(testCcsCred));
             });
 
             it("Caches token request correctly", async () => {
@@ -1039,8 +1140,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 expect(loginUrlSpy.calledWith(validatedRequest)).toBeTruthy();
             });
 
-            it(
-                "Does not use adal token from cache if it is present and SSO params have been given.",
+            it("Does not use adal token from cache if it is present and SSO params have been given.",
                 async () => {
                     const idTokenClaims: TokenClaims = {
                         "iss": "https://sts.windows.net/fa15d692-e9c7-4460-a743-29f2956fd429/",
@@ -1993,8 +2093,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             }
         });
 
-        it(
-            "Falls back to silent handler if thrown error is a refresh token expired error",
+        it("Falls back to silent handler if thrown error is a refresh token expired error",
             async () => {
                 const invalidGrantError: ServerError = new ServerError("invalid_grant", "AADSTS700081: The refresh token has expired due to maximum lifetime. The token was issued on xxxxxxx and the maximum allowed lifetime for this application is 1.00:00:00.\r\nTrace ID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxx\r\nCorrelation ID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxx\r\nTimestamp: 2020-0x-0x XX:XX:XXZ");
                 sinon.stub(RefreshTokenClient.prototype, <any>"acquireTokenByRefreshToken").rejects(invalidGrantError);
@@ -2074,6 +2173,170 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
                 expect(silentTokenHelperStub.calledWith(testNavUrl)).toBeTruthy();
             }
         );
+
+        it("makes one network request with multiple parallel silent requests with same request", async () => {
+            const testServerTokenResponse = {
+                token_type: TEST_CONFIG.TOKEN_TYPE_BEARER,
+                scope: TEST_CONFIG.DEFAULT_SCOPES.join(" "),
+                expires_in: TEST_TOKEN_LIFETIMES.DEFAULT_EXPIRES_IN,
+                ext_expires_in: TEST_TOKEN_LIFETIMES.DEFAULT_EXPIRES_IN,
+                access_token: TEST_TOKENS.ACCESS_TOKEN,
+                refresh_token: TEST_TOKENS.REFRESH_TOKEN,
+                id_token: TEST_TOKENS.IDTOKEN_V2
+            };
+            const testIdTokenClaims: TokenClaims = {
+                "ver": "2.0",
+                "iss": "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0",
+                "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
+                "name": "Abe Lincoln",
+                "preferred_username": "AbeLi@microsoft.com",
+                "oid": "00000000-0000-0000-66f3-3332eca7ea81",
+                "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
+                "nonce": "123523",
+            };
+            const testAccount: AccountInfo = {
+                homeAccountId: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
+                localAccountId: TEST_DATA_CLIENT_INFO.TEST_UID,
+                environment: "login.windows.net",
+                tenantId: testIdTokenClaims.tid || "",
+                username: testIdTokenClaims.preferred_username || ""
+            };
+            const testTokenResponse: AuthenticationResult = {
+                authority: TEST_CONFIG.validAuthority,
+                uniqueId: testIdTokenClaims.oid || "",
+                tenantId: testIdTokenClaims.tid || "",
+                scopes: [...TEST_CONFIG.DEFAULT_SCOPES, "User.Read"],
+                idToken: testServerTokenResponse.id_token,
+                idTokenClaims: testIdTokenClaims,
+                accessToken: testServerTokenResponse.access_token,
+                fromCache: false,
+                expiresOn: new Date(Date.now() + (testServerTokenResponse.expires_in * 1000)),
+                account: testAccount,
+                tokenType: AuthenticationScheme.BEARER
+            };
+            sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+            const silentATStub = sinon.stub(RefreshTokenClient.prototype, <any>"acquireTokenByRefreshToken").resolves(testTokenResponse);
+            const tokenRequest: CommonSilentFlowRequest = {
+                scopes: ["User.Read"],
+                account: testAccount,
+                authority: TEST_CONFIG.validAuthority,
+                authenticationScheme: AuthenticationScheme.BEARER,
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                forceRefresh: false
+            };
+            const expectedTokenRequest: CommonSilentFlowRequest = {
+                ...tokenRequest,
+                scopes: ["User.Read"],
+                authority: `${Constants.DEFAULT_AUTHORITY}`,
+                correlationId: RANDOM_TEST_GUID,
+                forceRefresh: false
+            };
+            const silentRequest1 = pca.acquireTokenSilent(tokenRequest);
+            const silentRequest2 = pca.acquireTokenSilent(tokenRequest);
+            const silentRequest3 = pca.acquireTokenSilent(tokenRequest);
+            const parallelResponse = await Promise.all([silentRequest1, silentRequest2, silentRequest3]);
+
+            expect(silentATStub.calledWith(expectedTokenRequest)).toBeTruthy();
+            expect(silentATStub.callCount).toEqual(1);
+            expect(parallelResponse[0]).toEqual(testTokenResponse);
+            expect(parallelResponse[1]).toEqual(testTokenResponse);
+            expect(parallelResponse[2]).toEqual(testTokenResponse);
+            expect(parallelResponse).toHaveLength(3);
+        });
+
+        it("makes network requests for each distinct request when acquireTokenSilent is called in parallel", async () => {
+            const testIdTokenClaims: TokenClaims = {
+                "ver": "2.0",
+                "iss": "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0",
+                "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
+                "name": "Abe Lincoln",
+                "preferred_username": "AbeLi@microsoft.com",
+                "oid": "00000000-0000-0000-66f3-3332eca7ea81",
+                "tid": "3338040d-6c67-4c5b-b112-36a304b66dad",
+                "nonce": "123523",
+            };
+            const testAccount: AccountInfo = {
+                homeAccountId: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
+                localAccountId: TEST_DATA_CLIENT_INFO.TEST_UID,
+                environment: "login.windows.net",
+                tenantId: testIdTokenClaims.tid || "",
+                username: testIdTokenClaims.preferred_username || ""
+            };
+            sinon.stub(CryptoOps.prototype, "createNewGuid").returns(RANDOM_TEST_GUID);
+            const silentATStub = sinon.stub(RefreshTokenClient.prototype, <any>"acquireTokenByRefreshToken");
+            const tokenRequest1: CommonSilentFlowRequest = {
+                scopes: ["User.Read"],
+                account: testAccount,
+                authority: TEST_CONFIG.validAuthority,
+                authenticationScheme: AuthenticationScheme.BEARER,
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                forceRefresh: false
+            };
+            const expectedTokenRequest1: CommonSilentFlowRequest = {
+                ...tokenRequest1,
+                scopes: ["User.Read"],
+                authority: `${Constants.DEFAULT_AUTHORITY}`,
+                correlationId: RANDOM_TEST_GUID,
+                forceRefresh: false
+            };
+            const tokenRequest2: CommonSilentFlowRequest = {
+                scopes: ["Mail.Read"],
+                account: testAccount,
+                authority: TEST_CONFIG.validAuthority,
+                authenticationScheme: AuthenticationScheme.BEARER,
+                correlationId: TEST_CONFIG.CORRELATION_ID,
+                forceRefresh: false
+            };
+            const expectedTokenRequest2: CommonSilentFlowRequest = {
+                ...tokenRequest1,
+                scopes: ["Mail.Read"],
+                authority: `${Constants.DEFAULT_AUTHORITY}`,
+                correlationId: RANDOM_TEST_GUID,
+                forceRefresh: false
+            };
+            const silentRequest1 = pca.acquireTokenSilent(tokenRequest1);
+            const silentRequest2 = pca.acquireTokenSilent(tokenRequest1);
+            const silentRequest3 = pca.acquireTokenSilent(tokenRequest2);
+            await Promise.all([silentRequest1, silentRequest2, silentRequest3]);
+
+            expect(silentATStub.calledWith(expectedTokenRequest1)).toBeTruthy();
+            expect(silentATStub.calledWith(expectedTokenRequest2)).toBeTruthy();
+            expect(silentATStub.callCount).toEqual(2);
+        });
+
+        it("throws error that SilentFlowClient.acquireToken() throws when making parallel requests", async () => {
+            const testError = {
+                errorCode: "create_login_url_error",
+                errorMessage: "Error in creating a login url"
+            };
+            const testAccount: AccountInfo = {
+                homeAccountId: TEST_DATA_CLIENT_INFO.TEST_HOME_ACCOUNT_ID,
+                localAccountId: TEST_DATA_CLIENT_INFO.TEST_UID,
+                environment: "login.windows.net",
+                tenantId: "testTenantId",
+                username: "username@contoso.com"
+            };
+            sinon.stub(RefreshTokenClient.prototype, <any>"acquireTokenByRefreshToken").rejects(testError);
+            try {
+                const tokenRequest = {
+                    scopes: TEST_CONFIG.DEFAULT_SCOPES,
+                    account: testAccount
+                };
+                const silentRequest1 = pca.acquireTokenSilent(tokenRequest);
+                const silentRequest2 = pca.acquireTokenSilent(tokenRequest);
+                const silentRequest3 = pca.acquireTokenSilent(tokenRequest);
+                await Promise.all([silentRequest1, silentRequest2, silentRequest3]);
+            } catch (e) {
+                // Test that error was cached for telemetry purposes and then thrown
+                expect(window.sessionStorage).toHaveLength(1);
+                const failures = window.sessionStorage.getItem(`server-telemetry-${TEST_CONFIG.MSAL_CLIENT_ID}`);
+                const failureObj = JSON.parse(failures || "") as ServerTelemetryEntity;
+                expect(failureObj.failedRequests).toHaveLength(2);
+                expect(failureObj.failedRequests[0]).toEqual(ApiId.acquireTokenSilent_silentFlow);
+                expect(failureObj.errors[0]).toEqual(testError.errorCode);
+                expect(e).toEqual(testError);
+            }
+        });
     });
 
     describe("logoutRedirect", () => {
@@ -2451,13 +2714,13 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
         it("getAccountByUsername returns null if account doesn't exist", () => {
             const account = pca.getAccountByUsername("this-email-doesnt-exist@microsoft.com");
-            expect(account).toBeNull;
+            expect(account).toBe(null);
         });
 
         it("getAccountByUsername returns null if passed username is null", () => {
             // @ts-ignore
             const account = pca.getAccountByUsername(null);
-            expect(account).toBeNull;
+            expect(account).toBe(null);
         });
 
         it("getAccountByHomeId returns account specified", () => {
@@ -2467,13 +2730,13 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
         it("getAccountByHomeId returns null if passed id doesn't exist", () => {
             const account = pca.getAccountByHomeId("this-id-doesnt-exist");
-            expect(account).toBeNull;
+            expect(account).toBe(null);
         });
 
         it("getAccountByHomeId returns null if passed id is null", () => {
             // @ts-ignore
             const account = pca.getAccountByHomeId(null);
-            expect(account).toBeNull;
+            expect(account).toBe(null);
         });
     });
 
@@ -2510,27 +2773,21 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
         it("active account is initialized as null", () => {
             // Public client should initialze with active account set to null.
-            expect(pca.getActiveAccount()).toBeNull;
+            expect(pca.getActiveAccount()).toBe(null);
         });
 
-        it(
-            "setActiveAccount() sets the active account local id value correctly",
-            () => {
-                expect((pca as any).activeLocalAccountId).toBeNull;
+        it("setActiveAccount() sets the active account local id value correctly", () => {
+                expect(pca.getActiveAccount()).toBe(null);
                 pca.setActiveAccount(testAccountInfo1);
-                expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
-            }
-        );
+                expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
+        });
 
-        it(
-            "getActiveAccount looks up the current account values and returns them()",
-            () => {
+        it("getActiveAccount looks up the current account values and returns them", () => {
                 pca.setActiveAccount(testAccountInfo1);
                 const activeAccount1 = pca.getActiveAccount();
                 expect(activeAccount1).toEqual(testAccountInfo1);
                 
                 const newName = "Ben Franklin";
-                window.sessionStorage.clear();
                 testAccountInfo1.name = newName;
                 testAccount1.name = newName;
                 const cacheKey = AccountEntity.generateAccountCacheKey(testAccountInfo1);
@@ -2538,8 +2795,7 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
 
                 const activeAccount2 = pca.getActiveAccount();
                 expect(activeAccount2).toEqual(testAccountInfo1);
-            }
-        );
+        });
 
         describe("activeAccount logout", () => {
             const testAccountInfo2: AccountInfo = {
@@ -2567,26 +2823,22 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             });
 
             it("Clears active account on logoutRedirect with no account", async () => {
-                expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+                expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                 await pca.logoutRedirect();
-                expect(pca.getActiveAccount()).toBeNull;
+                expect(pca.getActiveAccount()).toBe(null);
             });
     
-            it(
-                "Clears active account on logoutRedirect when the given account info matches",
-                async () => {
-                    expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+            it("Clears active account on logoutRedirect when the given account info matches", async () => {
+                    expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                     await pca.logoutRedirect({
                         account: testAccountInfo1
                     });
-                    expect(pca.getActiveAccount()).toBeNull;
+                    expect(pca.getActiveAccount()).toBe(null);
                 }
             );
 
-            it(
-                "Does not clear active account on logoutRedirect if given account object does not match",
-                async () => {
-                    expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+            it("Does not clear active account on logoutRedirect if given account object does not match", async () => {
+                    expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                     await pca.logoutRedirect({
                         account: testAccountInfo2
                     });
@@ -2595,26 +2847,22 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
             );
 
             it("Clears active account on logoutPopup with no account", async () => {
-                expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+                expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                 await pca.logoutPopup();
-                expect(pca.getActiveAccount()).toBeNull;
+                expect(pca.getActiveAccount()).toBe(null);
             });
     
-            it(
-                "Clears active account on logoutPopup when the given account info matches",
-                async () => {
-                    expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+            it("Clears active account on logoutPopup when the given account info matches", async () => {
+                    expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                     await pca.logoutPopup({
                         account: testAccountInfo1
                     });
-                    expect(pca.getActiveAccount()).toBeNull;
+                    expect(pca.getActiveAccount()).toBe(null);
                 }
             );
 
-            it(
-                "Does not clear active account on logoutPopup if given account object does not match",
-                async () => {
-                    expect((pca as any).activeLocalAccountId).toEqual(testAccountInfo1.localAccountId);
+            it("Does not clear active account on logoutPopup if given account object does not match", async () => {
+                    expect(pca.getActiveAccount()).toEqual(testAccountInfo1);
                     await pca.logoutPopup({
                         account: testAccountInfo2
                     });
@@ -2625,82 +2873,27 @@ describe("PublicClientApplication.ts Class Unit Tests", () => {
     });
 
     describe("Event API tests", () => {
-        it("can add an event callback and broadcast to it", (done) => {
+        it("can add an event callback", (done) => {
             const subscriber = (message: EventMessage) => {
                 expect(message.eventType).toEqual(EventType.LOGIN_START);
                 expect(message.interactionType).toEqual(InteractionType.Popup);
                 done();
             };
 
+            const callbackSpy = sinon.spy(EventHandler.prototype, "addEventCallback");
+
             pca.addEventCallback(subscriber);
-            // @ts-ignore
-            pca.emitEvent(EventType.LOGIN_START, InteractionType.Popup);
-        });
-
-        it("can remove an event callback", (done) => {
-            const subscriber = (message: EventMessage) => {
-                expect(message.eventType).toEqual(EventType.LOGIN_START);
-                expect(message.interactionType).toEqual(InteractionType.Popup);
-            };
-
-            const callbackSpy = sinon.spy(subscriber);
-
-            const callbackId = pca.addEventCallback(callbackSpy);
-            // @ts-ignore
-            pca.emitEvent(EventType.LOGIN_START, InteractionType.Popup);
-            pca.removeEventCallback(callbackId || "");
-            // @ts-ignore
-            pca.emitEvent(EventType.LOGIN_START, InteractionType.Popup);
             expect(callbackSpy.calledOnce).toBeTruthy();
             done();
         });
 
-        it("can add multiple callbacks and broadcast to all", (done) => {
-            const subscriber1 = (message: EventMessage) => {
-                expect(message.eventType).toEqual(EventType.ACQUIRE_TOKEN_START);
-                expect(message.interactionType).toEqual(InteractionType.Redirect);
-            };
+        it("can remove an event callback", (done) => {
+            const callbackSpy = sinon.spy(EventHandler.prototype, "removeEventCallback");
 
-            const subscriber2 = (message: EventMessage) => {
-                expect(message.eventType).toEqual(EventType.ACQUIRE_TOKEN_START);
-                expect(message.interactionType).toEqual(InteractionType.Redirect);
-                done();
-            };
-
-            pca.addEventCallback(subscriber1);
-            pca.addEventCallback(subscriber2);
-            // @ts-ignore
-            pca.emitEvent(EventType.ACQUIRE_TOKEN_START, InteractionType.Redirect);
-        });
-
-        it("sets interactionType, payload, and error to null by default", (done) => {
-            const subscriber = (message: EventMessage) => {
-                expect(message.eventType).toEqual(EventType.LOGIN_START);
-                expect(message.interactionType).toBeNull();
-                expect(message.payload).toBeNull();
-                expect(message.error).toBeNull();
-                expect(message.timestamp).not.toBeNull();
-                done();
-            };
-
-            pca.addEventCallback(subscriber);
-            // @ts-ignore
-            pca.emitEvent(EventType.LOGIN_START);
-        });
-
-        it("sets all expected fields on event", (done) => {
-            const subscriber = (message: EventMessage) => {
-                expect(message.eventType).toEqual(EventType.LOGIN_START);
-                expect(message.interactionType).toEqual(InteractionType.Silent);
-                expect(message.payload).toEqual({scopes: ["user.read"]});
-                expect(message.error).toBeNull();
-                expect(message.timestamp).not.toBeNull();
-                done();
-            };
-
-            pca.addEventCallback(subscriber);
-            // @ts-ignore
-            pca.emitEvent(EventType.LOGIN_START, InteractionType.Silent, {scopes: ["user.read"]}, null);
+            const callbackId = pca.addEventCallback(() => {});
+            pca.removeEventCallback(callbackId || "");
+            expect(callbackSpy.calledOnce).toBeTruthy();
+            done();
         });
     });